refactor: moves all signup form components to one place
