+++ src/views/certifications/views/landing/components/signup-form-area/components/newsletter-signup-form/marketo-form-data.ts
