+++ src/views/certifications/views/landing/components/signup-form-area/components/signup-header/index.tsx
